Implement authentication using Google sign-in and Firebase (#33)
# Title:
Implement authentication using Google sign-in and Firebase

# Description:
## Purpose of the change
This PR implements the Authentication mechanisms using Google and
Firebase to sign in. It now allows a user to sign in to wildex using
their google account.
It also includes a SignInScreen, simple placeholder screen to test
manually if one can sign in successfully.

Note: 
- Update the gitignore to include some firebase files
- Update build.gradle.kts and libs.versions.toml to incluse missing
dependecies
- Portions of this code are copy-pasted from the Bootcamp solution
provided by the SwEnt staff
- The methods in AuthRepositoryFirebase are marked as TODO() as they are
taken care of in issue #77 because this PR was too big.

## Related issues
#21 

## How to test
1. Configure Firebase in your local environment and ensure the Google
Sign-In OAuth client is set up correctly.
2. Verify the Emulator is started.
3. Run the tests.