feat: Implement HomeScreen UI (#28)
## Description
This PR implements the first version of `HomeScreen` with its
corresponding tests `HomeScreenTest`. It also adds the necessary
resources to ensure proper display of all UI elements.

## Related issues
Closes #18 

## Trade-offs or known limitations
Some issues occur when fully integrating the ViewModel, causing the app
to crash. The root cause is still unknown so the screen uses default
values until the issue is resolved.

## How to test
1. Run `HomeScreenTest.kt`

---------

Co-authored-by: Kozi <steph4140@gmail.com>
Co-authored-by: Martin Zimmer <mzimmer@ic2-adrstd-5022.intranet.epfl.ch>
Co-authored-by: Martin Zimmer <mzimmer@ic2-adrstd-5002.intranet.epfl.ch>
Co-authored-by: Martin Zimmer <mzimmer@ic2-adrstd-5008.intranet.epfl.ch>