refactor: cleanup home screen UI design and functionnality code (#103)
- Introduced local in-memory repositories for testing
- Minor corrections to Firestore-related constants and test
data.
- Added & adapted ui tests to home screen and unit tests to its view model
- Updated the Firestore users collection path constant from `"Users"` to
`"users"` in both production and test code for consistency.
- Updated test post data in `FirestoreTest` to use realistic image URLs
for `pictureURL` fields, improving test fidelity.
- Registered a singleton `LikeRepository` in the `RepositoryProvider`
for use throughout the app.

## Related issues
Advances on #79 

## Trade-offs or known limitations 
- None at this time

## How to test
* Modified `MainActivity` to perform anonymous Firebase authentication
before setting the app content, and updated the app entry point to use
`HomeScreen` directly.
* Since the real firestore is populated with posts, all that is to do is
run the MainActivity, the Home screen is already called on `WildexApp`.
It simply has to be commented out along with the anonymous sign in line.