feat: add animal information UI and VM skeleton (#170)
## Description
This pull request introduces a new animal information screen to the app
by implementing a basic skeleton for the screen and the view model.
Following the current figma design, this screen is to be accessed from
the collection screen, to which one can go back. The idea is to have a
larger image of the animal displayed with all the corresponding
information such as the name, species and description.

- Added `AnimalInformationScreen.kt` and
`AnimalInformationScreenViewModel.kt`

## Related issues
Closes #100 

## Trade-offs or known limitations 
-Still needs implementation of the animal repository