feat: implement Collection screen VVM skeleton
I added the skeleton of the view and the view model of the
collection screen. There are more arguments to the collection screen
than needed for any version of the view but all are needed so that we can
display the collection of the current user or of any other user.