Implement Collection Screen View and ViewModel skeleton
## Description
This PR adds the skeleton of the View and of the ViewModel of the
`CollectionScreen`. The View will need to support two different versions
depending on whether we want to display the current user's collection or
that of any other user.
If we want to display the current user's collection, we need to display
:
- Bottom bar -> `bottomBar` is needed in the screen
- Profile button -> `onProfileClick` is needed for that
- Notifications button -> `onNotificationClick` is needed

Otherwise we need to display only the Go Back button -> `onGoBack` is
needed.

## Related issues
Closes #98 

## Trade-offs or known limitations 
The interfaces of the `UserAnimals` and `Animal` repositories are needed
to define the `CollectionScreenViewModel` any further.