feat: implement animal repository (#175)
## Description
This pull request introduces an animal repository by adding an interface
and a firestore implementation. The repository is necessary for view
models like those of the animal information, the collection, the post
details and the post creation. Unlike some of the other repositories, we
do not use a `getNewAnimalId` method from firestore to create the id of
an item, but rather use that which will be created from the animal
detetct database.

Note: Some refactoring changes were necessary to avoid confusion with
regards to the `AnimalRepository` for the `Animal` class and the
`AnimalInfoRepository` for the `AnimalDetect` functionality.

-Added `AnimalRepository.kt`, `AnimalRepositoryFirestore.kt` and
`AnimalRepositoryFirestoreTest.kt`
-Refactored old `AnimalDetectRepository.kt` (that temporarly was
`AnimalRepository.kt`) to `AnimalInfoRepository.kt`
-Modified several files to include these new changes

## Related issues
Closes #117 

## Trade-offs or known limitations
-No `editAnimal` or `deleteAnimal` methods are implemented, as this
doesn't make sense for now. This may change in the future

## How to Test
Laucnh firebase emualtors (only firestore necessary) and run
`AnimalRepositoryFirestoreTest.kt`