feat: add animal list view and integrate to nav
I added the view for the animal list, the column as well as each
individual animal's card.
I also integrated the collection screen into the main activity and into
the navigation actions so that it is accessible to view the current
user's collection as well as any other user's collection.