Feature/93 edit profile viewmodel (#180)
## Description
This PR implements the EditProfile viewModel. It fetches all current
profiles informations on repositories and gives them to the screen to be
displayed and modified. Once done, it takes the new edited infos from
the screen and sends uploads them to their respective repositories.

## Related issues
Closes #93

## How to test :
Run EditProfileViewModelTest
## Note :
When working on the Edit screen, the viewModel might still change
accordingly if we decide to implement a new feature in the screen.