test: add first tests
I wrote the tests for the empty collection as well as the tests to
differentiate the collection of the current user or that of another
user. I also wrote the tests for the callback buttons. The collection
list itself and the callback for animal click are yet to be tested.