Implement Collection Screen View and ViewModel
## Description
This PR implements the `CollectionScreen.kt` as well as the
`CollectionScreenViewModel.kt`. This allows the user of the app to
access his own collection of animals, but also access any user's
collection with a click from its profile. By viewing a user's
collection, we mean seeing the unlocked animals and allowing to click on
them to be redirected to the `AnimalDetailsScreen` and also see the yet
to be discovered animals. These locked animals cannot be clicked on and
a user needs to first unlock them to access their details.

## Related issues
Closes #75 and #99 

## Known limitations
-The link from the `CollectionScreen` to the `AnimalDetailsScreen` is
yet to be implemented

## How to test
-Run `CollectionScreenTest.kt`