feat: implement EditProfileScreen (#68) (#205)
## Description
This PR implements the EditProfile screen and updates its viewModel
adding useful functions. It includes tests for both files.
The screen purpose is to modify the current user's informations, and
save them in the online repo. They can change their names, bio, country
and profile picture.

* Add ability to access the phone's pictures to upload a new profile
picture.
* Text fields display an error when upload conditions aren't met.
* The save button uploads all modifications to the online repository and
depending on wether it's the first time signin in or not, will bring
back to the home screen or stay there if it's not a new user.
* When the screen loads, the text fields are already filled with the
user's infos pulled from repository.
* The screen is scrollable to adapt to any screen size or shape.
* It shows a loading screen when loading and an error screen if
something went wrong.

## Related issues
Closes #68 

## Trade-offs or known limitations
Currently, navigation doesn't take this screen into account, so it is
not accessible yet.

## How to test
* Run EditProfileScreenTest
* (To check updated tests) Run EditProfileViewModelTest

## Notes:
* AI assistant was used to help debugging tests