fix: Refactor the comment data class and repository (#207)
## Description
This PR adds two functions to the `CommentsRepository` interface, their
implementation in `CommentsRepositoryFirestore` along with new tests in
`CommentsRepositoryFirestoreTest`.

These new functions provide the following operations:
- deleteAllCommentsOfPost(postId): Deletes all comments of a post
- deleteAllCommentsOfReport(reportId): Deletes all comments of a report

This required to add a tag to the `Comment` data class, specifying if a
comment is related to a `Post` or a `Report`, so that we can
differentiate a `Post` with a `Report` in case they have the same Id -
which is possible since they come from different collections.

This implied necessary changes in every file related to Comments, so
they could handle this new tag.

Some changes were also made to use the alias Id instead of String, for
better clarity.

## Related issues
Closes #196 

## How to test
1. Start the Firestore Emulator: `firebase emulators:start`
2. Run `CommentRepositoryFirestoreTest.kt`

---------

Co-authored-by: Martin Zimmer <mzimmer@ic2-adrstd-5009.intranet.epfl.ch>