feat: implement report screen view model and refactor (#208)
## Description
This PR introduces the `ReportScreenViewModel`, responsible for managing
and exposing the UI state of the `ReportScreen`. It handles loading,
refreshing, and performing actions on reports based on the user type,
while also refactoring the Report data model to simplify its structure.

### Details for the `ReportScreenViewModel`
- It loads reports based on the user type. `PROFESSIONAL` users can see
all reports, while `REGULAR` users only see their own.
- It maintains a clear UI state object containing all necessary
information, and includes `loadUIState()` and `refreshUIState()`
functions to handle state transitions.
- It implements all report actions : 
- `cancelReport()` for when any user wants to remove the report they
made
- `selfAssignReport()` for when a `PROFESSIONAL` user wants to assign a
report to themselves
- `unselfAssignReport()` for when a `PROFESSIONAL` user wants to
unassign themselves from a report they had previously had self-assigned
to
- `resolveReport()` for when a `PROFESSIONAL` user has resolved a report
(and thus wants to remove it)

### Details for the `Report` refactoring
- The previous field `status` in the data class is currently not needed,
as the information the UI needs to correctly display the report status
can be computed from the `assigneeId` field of the `Report` data class.
- Removed the `ReportStatus` enum
- Cleaned up `ReportRepositoryFirestore`,
`ReportRepositoryFirestoreTest`, `FirestoreTest`, `Map`,
`DataClassesTest`, `MapScreenViewModel` and `MapScreenViewModelTest` to
account for this change


## Related issues
Closes #195 

## Trade-offs or known limitations 
- The `updateUIState()` function both loads the `User` and `SimpleUser`
instances of the current user in order to get its user type, since this
field is not included in the `SimpleUser` data class. These repository
accesses might be redundant, and it may be better to just load the
`User` and create a `SimpleUser` instance based on the loaded `User`
instance.

## How to test
Run `ReportScreenViewModelTest`