feat: implement Settings screen ViewModel
I implemented the Settings screen ViewModel. In order to implement
it, I needed to create a `UserSettings` data class which stores each
user's preferences. For the moment it only contains the notification
preference and the app theme preference.
I also wrote a `UserSettingsRepository` interface and a
`UserSettingsRepositoryFirestore` implementation in order to store and
retrieve these settings from the ViewModel. It allows the basic
operations of setting and getting specific users' preferences.
Then I implemented the `SettingsScreenViewModel` which defines all the
available operations in the settings, including changing the app theme,
enabling notifications and changing the user status. This ViewModel
relies on a `UserSettingsRepository` to store and get the notification
status and the app theme, and on a `UserRepository` to store and get the
user's status.