feat: add user animals repo deletion function
I added a function to the user animals repo that allows to delete the
animals linked to the given user so that we can easily delete a user's
account. I also wrote the tests to verify the proper functionality of
the function. I added a use case class that handles the deletion of a
user's account