feat: implement Settings screen
I implemented the settings screen which allows to manage the
currently logged in user's preferences like app appearance mode, the
user's status or the notifications enablement. It also allows the user
to delete his account or sign out from the app. To avoid involuntary
account deletions, when a user wants to delete his account, a
confirmation popup appears to make sure the operation was initiated on
purpose. Once a user deletes his account, he is taken back to the sign
in screen where he can create a new account if he wishes.

The theme of the app changes dynamically as the user
changes the appearance mode.

In order to implement the "Delete account" button, I had to implement
some new functions in the existing repositories. These new functions
are:
- `deleteUserSettings(userId: Id)` in `UserSettingsRepository`
- `deleteUserAnimals(userId: Id)` in `UserAnimalsRepository`
- `deleteUserAchievements(userId: Id)` in `UserAchievementsRepository`
- `deletePostsByUser(userId: Id)` in `PostsRepository`
- `deleteReportsByUser(userId: Id)` in `ReportRepository`
- `deleteLikesByUser(userId: Id)` in `LikeRepository`
- `deleteCommentsByUser(userId: Id)` in `CommentRepository`
These added functions have been tested and their definitions have been
added to every implementation of the repositories' interfaces.

With these newly added functions, I was able to define the use case
`DeleteUserUseCase` which handles deleting a user's account from the
database. While I was at it, I also defined another use case
`InitializeUserUseCase` which handles creating a user's account on the
first sign in. This creation use case was integrated into the
`SignInViewModel`.

Finally, I deleted the function `getNewUid()` from the `UserRepository`
as it was never used and the identifiers of the user are taken from the
Firebase user identifiers.