feat: submit report screen
## Description
This PR introduces the implementation of the `SubmitReportScreen` which
allows the user to report a wounded animal or an animal in need of
rescue.
To do so, one must take a picture of the animal and provide a
description before submitting. As the end goal is to signal to
professionals that an animal needs their help, the user has to provide a
precise location, making camera and location access mandatory.
Nonetheless permission is asked explicitly and the user has the liberty
to refuse, in which case they can not submit their report.

Here are some screenshots of how this screen works.
At first, the `SubmitReportFormScreen` is displayed, allowing the user
to upload a picture, they can either take one or upload one from
gallery, they also must provide a description.
When providing a picture, if the camera permission access is confirmed,
the `CameraPreviewScreen` is displayed. If not, the
`CameraPermissionScreen` prompts the user to accept the camera access
request, explaining what it's needed for.
Before submitting, if location permission access is not granted, a pop
requests it, if accepted and location is successfully fetched, then the
user can submit their report.
<table>
  <tr>
<td><img width="630" height="1305" alt="image"
src="https://github.com/user-attachments/assets/be4f094f-03eb-4c7a-8484-ebcc6184db74"
/></td>
<td><img width="630" height="1305" alt="image"
src="https://github.com/user-attachments/assets/9ab4ad40-1da2-4ae4-94a9-51dbd6596c97"
/></td>
  </tr>
  <tr>
<td><img width="630" height="1305" alt="image"
src="https://github.com/user-attachments/assets/6b470490-0643-4ff0-8773-8f5b56b7bcc2"
/></td>
<td><img width="630" height="1305" alt="image"
src="https://github.com/user-attachments/assets/d5b23910-e302-4512-bce6-ca441bd3456d"
/></td>
  </tr>
</table>

## Related issue
Closes #222 

## Trade-off or known limitations
- Camera and location permission are necessary. Camera access can be
denied if the user prefers to upload a picture, but they must provide a
picture one way or another.
- There is no work around for location access. We could maybe provide a
way to manually select the location in the future.

## How to test
- Launch the app and navigate to the SumitReportScreen and try to submit
a report
- Run SubmitReportFormScreenTest and SubmitReportScreenTest