feat: Implement UserFriends data class and its repository interface (#270)
## Description
This PR defines the `UserFriends` data class, which is used to keep
track of a user’s friends.

It also implements the `UserFriendsRepository` interface, providing
functions that can:
- **Initialize** the `UserFriends` of a new user with an empty
*friendIds* list and a *friendsCount* of 0.
- **Retrieve** the *friendIds* or *friendsCount* from a user’s
`UserFriends`.
- **Add** a friend to a user’s `UserFriends`.
- **Remove** a friend from a user’s `UserFriends`.
- **Delete** a user’s `UserFriends`.

## Related issues
Closes #262 

## Trade-offs or known limitations 
This comment isn’t specific to `UserFriends`, but applies to all
user-related data classes and repositories (`UserAnimals`,
`UserAchievements` and `UserFriends`).

I noticed that the function retrieving the list of animals (or
achievements) in `UserAnimals` (or `UserAchievements`) literally returns
a list of `Animal` (or `Achievement`) objects. An approach that might
not be optimal since these lists could be large without always needing
all their data.

I think it would be better to only return the list of IDs of the
`Animal` (or `Achievement`) objects, and then fetch the full details
using the corresponding repository if needed. At least that's what I did
for `UserFriends`, and I think we should do the same for the other
classes and repositories (in a separate PR).

Co-authored-by: Martin Zimmer <mzimmer@ic2-adrstd-5006.intranet.epfl.ch>