test: add tests for UserRecommender + fix issues
I added tests for the UserRecommender. I also fixed an issue with the
distance computing. There was a problem with the magnitude of the
distance and thus most scores would get rounded to 0 for the
geographical proximity. Now each user is put on a linear scale from the
minimum distance to the maximum distance from any user to the current
user.