fix: fix close by factor compute
I fixed the computing for the close by factor. Before, distances were
put on a linear scale from the minDistance to the maxDistance. Now, they
are put on a linear scale from 0 to the maxDistance. This prevents cases
where all users are very far away to have a very far away user have a
very high close by factor.