refactor: add user typer to simple user (#281)
## Description
This PR refactors the `SimpleUser` data class to include a `userType`
field. This change is needed to facilitate the #265 issue: in screens
where we use `SimpleUser` instances to display clickable profile
pictures, the user type will determine whether the professional badge
will be displayed. It will also allow some screens to have all
information needed through just `getSimpleUser()`, instead of fetching a
lot of extra unnecessary information with `getUser()` (as shown in the
refactoring of `ReportScreenViewModel`)

The `SimpleUser` data class is of course mofidifed, and several view
models and tests are updated in order to reflect this change. It's
mostly correcting instances of `SimpleUser`

## Related issues
Closes #280 

## Trade-offs or known limitations 
Some other screens may also benefit from a logic refactoring where this
new refactoring is utilized. We should go through and check that during
the next meeting.