feat: implement friend suggestion algorithm
I implemented a friend suggestion algorithm to encourage new
connections between users within the app. The suggestion of a user to
the current user of the app is computed based on three parameters that
illustrate how attractive a profile is to the current user. These
parameters are the following:
- **Mutual friends:** a user who shares many common friends with the
current user is a good candidate for a new connection
- **Popularity:** a user who has many friends is likely very active and
also makes good posts, making them a good candidate for a new connection
- **Close by activity:** a user who posts regularly, close to the
current user's location is a good candidate for a new connection

The rank of a user in the suggestion list is defined based on a weighted
sum of these three parameters. The order of importance of the parameters
in the rank is the following:
1. Mutual friends (50%)
2. Close by activity (30%)
3. Popularity (20%)

The suggestion algorithm returns a list of relevant users, each
accompanied with a suggestion reason which represents the main reason
behind the user being suggested. This reason is chosen based on the
highest contribution of all parameters to the user's rank. By
contribution, understand the parameter's score (between 0 and 1)
multiplied by its weight in the rank's weighted sum. This reason is
meant to be displayed for the current user to see, in order to get a
better sense of why a suggested user should be of interest to them.

In order to implement this algorithm, a new function `getAllUsers()`
needed to be added to the `UserRepository` in order to get all the users
from the database for the suggestion score computing. Two new
`LocalRepositories` were added to the project for testing purposes:
- `UserFriendsRepository`
- `RelationshipRepository`