feat: add VM tests and modify friendState
I added the tests for the view model and I modified the FriendState data
class to add another boolean reflecting wether this friend state has as
its user the current user. This allows to not display any interactable
when the current user is in the friend list of another user and they
view themself.