feat: implement notification screen (#248)
## Description
This PR implements the notification screen and its tests. It displays
them with a specific description depending on the notification type. It
redirects the user to the corresponding screen when they click on the
right arrow. If they click on the profile picture of the notification's
author they get redirected to the author's profile.

## Screen's UI overview
<img width="191" height="266" alt="image"
src="https://github.com/user-attachments/assets/5b5a9c3b-ac42-424f-a781-8f73bb5a93fb"
/>


## Related issues
Closes #248 

## How to test
1. Run NotificationScreenTest.kt

## Trade-offs or known limitations
* As notifications are still being implemented by another team member, I
can't fully implement the viewModel first so I started working on the UI
and will fix its behaviors once the viewModel is done (which implies:
Once the notification class is created by my team mate)
* Currently as the viewModel is not completely implemented yet, loading
features are either commented or will load infinitely (the viewModel
needs to reset this state)
* Navigation currently doesn't take into account the path to this screen

## Notes
* This PR required a partial implementation of the viewModel, it was not
intended to have a fully working one now, it is a job for a futur PR to
implement the viewModel specifically