feat: implement Friend Screen ViewModel
I implemented the ViewModel of the Friend Screen. The operations
allowed by this ViewModel are the following:
- Send a friend request
- Cancel a sent friend request
- Accept a received friend request
- Decline a received friend request
- Unfollow a user

All of these operations are implemented using a rollback mechanism so
that latency for screen updates is avoided but state integrity is still
maintained by the rollback in case of an exception in the repositories
(inspiration was taken from @Rania5724 's Report Details Screen
ViewModel).

The ViewModel includes support for a loading screen and also refreshing.

About suggestions from the friend suggestion algorithm, the ViewModel
maintains at all times a list of suggested users. Whenever a user from
the suggestions is followed, a new suggestion replaces the old one (if
there are still users to suggest) to always have some suggestions for
the current user. To limit computations, suggestions are recomputed only
when the maintained list becomes too small, or when the screen is
refreshed.