feature: improve achievement screen (#293)
## Description
This PR refactors the Achievements screen architecture and introduces a
new interactive feature: when the user clicks on an achievement, a
dialog window displays information on the achievement: name, picture and
description.
The screen architecture has been cleaned up to improve readability.
This includes an update to the navigation in `MainActivity` as the
`AchievementsScreen` now requires a `userId`.

<table>
  <tr>
<td><img width="621" height="1293" alt="image"
src="https://github.com/user-attachments/assets/787f9236-ffa2-46f7-bae2-25cc4169ad60"
/></td>
<td><img width="621" height="1293" alt="image"
src="https://github.com/user-attachments/assets/ca62ebc3-2868-4e95-a1c3-204af682e9aa"
/></td>
  </tr>
</table>

## Related issue
Closes #264 

## Trade-offs or known limitations
- The details dialog currently displays basic achievement information,
support for images or richer content may be added later.
- The `navigation_AchievementsScreenFromProfile`() fails due to changes
of parameters. This is commented for now, but needs to be fixed later
on.

## How to test
1. Run AchievementsScreenTest
2. Run AchievementsScreenViewModelTest