feat: implement Friend Screen View
I implemented the Friend Screen View. The Friend Screen allows the
logged in user to access their social menu where they can manage their
friends, see new friend suggestions and manage their friend requests
(sent or received). In their social menu, the logged in user can
navigate between two tabs, one for friends management and the other for
friend requests management. 
The Friend Screen also allows the logged in user to take a look at other
users' friend lists and manage their status from there, so that the
current user can follow friends of a user whose friend list they are
viewing. When a specific section of the friend screen is empty (friend
list, suggestion list, received requests list or sent requests list), a
placeholder is displayed to show it to the user. 

Suggestions take a while to load so to avoid having the screen marked
as loading as long as the suggestions are being computed, I made their
computing asynchronous and independent of the screen's loading state.
This means the suggestions are shown as empty as long as they are being
computed.