feat: implement and test offline screen (#303)
## Description
This PR introduces the implementation of an `OfflineScreen` that will be
displayed in case a certain functionality or screen is not available
when the user is offline.
It also includes a fairly basic test of the implementation.
I decided to put this screen in the utils package as it will be used by
many screens in the following changes for the offline implementation.

## UI preview
<img width="270" height="600" alt="image"
src="https://github.com/user-attachments/assets/68a6f11f-c3b6-4f9e-b09a-b497c5628e49"
/>

## Related issue
Related to #297 

## Trade-offs or known limitations
None for now.

## How to test
As this screen is not yet implemented in the navigation, the only way to
test it is to run `OfflineScreenTest`