feat: create search bar composable and auto update
I added the definition fro the search bar composable which I integrated
into the friend screen. I also added an auto-update mechanism in the
UserIndex so that whenever the local data file is written to, the map
used for the index is updated as well to reflect the latest changes.