feat: implement user search bar (#330)
I implemented a search bar that allows fast user searching on
Wildex. This search bar relies on multiple components and I'll go
through them from the lowest abstraction level to the highest.

### `FileSearchDataStorage`
This class is responsible for maintaining a local file containing all
the data necessary to the user search. This file is placed in the cache
directory of the Android device's environment so that it gets deleted
after an app kill. The idea of this file is to temporarily exist during
the lifetime of the app so that it allows fast user searching by
reducing Firestore database queries. The content of the file is a Json
string that represents a map from a user's string representation for the
search algorithm to this user's id. This class allows writing and
reading to the file and also has a status for the file data, a boolean
which is true when the file was written to but not yet read, false
otherwise. This allows reading only when needed as one should read the
file only if they haven't read it yet or if the file was updated.

### `SearchDataUpdater`
This class is responsible for updating the search data by accessing the
user repository and writing to the local file the latest version of the
database. After fetching all users from the database, the updater
defines the map to store in the file. This map is a mapping from the
users' string representations to the users' identifiers. A user's string
representation is of the following format: `<user's name> <user's
surname> <user's username>`. I will explain why this format was chosen
later. The reason why we map this string representation to the user's
identifier is because if a query matches a user's string representation,
we can easily access the user's id from there, and then access their
profile screen, display the result with their profile picture, name,
surname, etc.

### `SearchDataProvider`
This class is responsible for providing the search data by reading the
local file via the file maintainer and caching the search data. This
provider caches the search data and returns it whenever it is asked, but
also exposes the file data's status by broadcasting the boolean file
status from the file maintainer. When this boolean is true, one can
invalidate the cache so that the new search data is fetched and cached.

### `UserIndex`
This class is responsible for matching search queries to users present
in the search data. The algorithm works as follows:
1. The search query is separated into subqueries, the separator being
the space character. For example, the query `leo lem` is divided into
two subqueries `leo` and `lem`.
2. Each subquery is mapped to a Pattern which, for example, allows the
user to get results of users with accented letters in their names
without having to write the accent in the search query. So, each letter
with accented versions is mapped to the set of its accented versions
plus itself. `a -> [aáàâä]`. Thanks to this, typing `leo` will match
with `léo`. Also, queries which contain only lowercase letters are
matched with strings, insensitively of the case so that for example
`leo` is matched with `Leo`. But if the user makes the effort to write
an uppercase letter, we consider that they don't want to match with
lowercase letters. This means that typing `A` will only match strings
like `Arthur` or `Andy` but not `Paul` even though it contains an `a`.
3. Each user's string representation is given a score relative to the
search subqueries, equal to the sum of the representation score relative
to each subquery. As a reminder, a user's string representation for the
search algorithm is `<user's name> <user's surname> <user's username>`.
This minimises users' representation collisions (see Known Limitations).
The score of a user's string representation relative to a subquery is
computed as follows: the representation is given a base score equal to
`100 * (length of the matching between the subquery and the
representation / the length of the representation)`. Then, if the
subquery matches with the beginning of any component of the user's
representation, its base score is multiplied by 4. Finally, if the
subquery matches with the end of any component of the user's
representation, its base score is multiplied by 2. **For example**, let
`Léonard Lemaire leolmr1110` be a user's string representation and
`leonard` a subquery. The base score of this representation relative to
the subquery is: `100 * (7 / 26)`. Then since the subquery matches with
the beginning of `Léonard`, the base score is multiplied by 4. Then, the
subquery also matches with the end of `Léonard` thus the base score is
multiplied by 2, yielding a final base score multiplicator of 8.
Finally, the score of this representation relative to this subquery is:
`8 * 100 * (7 / 26) = 215,38`.
4. Users' string representations are sorted by descending score relative
to the search query and mapped to their corresponding `User`.

### `UserSearchBar`
This Composable defines the appearance of the user search bar as well as
the UI logic related to it. The search bar was integrated into the
Friend screen. I used the predefined Composable `SearchBar`. When the
user focuses on the input field, the search bar expands into a search
view that takes all available space within the parent. The leading icon
of the search bar allows to leave this search view to return to the
default Friend screen view. The trailing icon of the search bar clears
any text in the input field. Search results are displayed in a
`LazyColumn` below the search bar and clicking on a search result
navigates the current user to the result's profile screen.

> [!NOTE]
> The search results algorithm wasn't designed by me, I adapted it to
user searching, taking inspiration from the work of professor Michel
Schinz, who wrote it for the CS-108 project (ReChor) from last year for
public transportation stops searching.

## Trade-offs or known limitations
- Since username uniqueness isn't yet enforced on the app, there is a
bug whenever two users have exactly the same name, surname and username.
Having this collision is unlikely for a medium sized app with a limited
number of users but this can happen very quickly as the number of users
grows.
- There is one SonarCloud issue which I didn't fix because the solution
doesn't work as I want it to. The non-deprecated version of the
`SearchBarColors` constructor takes one more argument than what I
passed, an additional `inputFieldColors`. However, when passing the
`InputFieldColors` to the `SearchBarColors` instead of to the `colors`
argument of the `InputField`, the passed colours are simply ignored and
not applied to the input field. Thus, I decided to leave it as is so
that I can apply colours to the input field and accept this issue as
it's minor.

## How to test
- Run `FriendScreenTest`
- Run `UserIndexTest`
- Run `SearchDataProviderTest`
- Run `SearchDataUpdaterTest`
- Run `FileSearchDataStorageTest`
- Navigate to the Friend screen and try different inputs to see if the
results are relevant