feat: improve Home Screen View (#339)
I improved the design of the Home Screen and fixed the appearance
mode of the app. The appearance mode was previously only set in the
`SignInViewModel` but that introduced a bug where if a signed in user
closed the app, and re-opened it, the appearance mode would be set to
follow the system's appearance mode instead of the user's appearance
mode defined in their settings. Now, the appearance mode is set in the
`HomeScreenViewModel`.
The UI improvements for the Home Screen consist of the following:
- The posts' border was removed
- The posts' width was increased to take up all screen's width
- The posts' location is now centered vertically relative to the header
row, as opposed to before where it was aligned with the post's date
- The images' height was increased
- The like button on the post's image was removed because it didn't look
good dependent on the image, and it was redundant with the like button
under the image.
- The actions icons' sizes were increased
- The posts' title was changed -> animals' names are highlighted and
capitalised

## How to test
- Run `HomeScreenTest`
- Run `HomeScreenViewModelTest`
- Navigate to the Home Screen and see changes live